+++ src/dbconnection.js
@@ -1,0 +1,23 @@
+const mysql = require('mysql');
+
+// Create MySQL connection
+const connection = mysql.createConnection({
+    host: 'localhost',
+    user: 'root',
+    password: '',
+    database: 'learning_management'
+});
+
+
+function connectdb() {
+    // Connect to MySQL
+    connection.connect((err) => {
+        if (err) {
+            console.error('Error connecting to MySQL: ' + err.stack);
+            return;
+        }
+        console.log('Connected to MySQL as id ' + connection.threadId);
+    });
+}
+
+module.exports = connectdb;
