+++ src/html/index.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Learning Management App</title>
+    <link rel="stylesheet" href="src/styles/style.css">
+</head>
+<body>
+    <h1>Learning Management App</h1>
+    <form id="register-form">
+        <h2>Register</h2>
+        <input type="text" name="username" placeholder="Username" required>
+        <input type="password" name="password" placeholder="Password" required>
+        <input type="email" name="email" placeholder="Email" required>
+        <input type="text" name="full_name" placeholder="Full Name" required>
+        <button type="submit">Register</button>
+    </form>
+    <form id="login-form">
+        <h2>Login</h2>
+        <input type="text" name="username" placeholder="Username" required>
+        <input type="password" name="password" placeholder="Password" required>
+        <button type="submit">Login</button>
+    </form>
+    <form id="logout-form">
+        <button type="submit">Logout</button>
+    </form>
+    <script src="src/js/script.js"></script>
+</body>
+</html>
