+++ src/routes.js
@@ -1,0 +1,127 @@
+const express  = require('express');
+const route = express.Router();
+const { check, validationResult } = require('express-validator');
+const bcrypt = require('bcryptjs');
+
+
+route.get('/', (req, res) => {
+    res.sendFile(__dirname + '/index.html');
+});
+
+  
+// Define a User representation for clarity
+const User = {
+    tableName: 'users', 
+    createUser: function(newUser, callback) {
+        connection.query('INSERT INTO ' + this.tableName + ' SET ?', newUser, callback);
+    },  
+    getUserByEmail: function(email, callback) {
+        connection.query('SELECT * FROM ' + this.tableName + ' WHERE email = ?', email, callback);
+    },
+    getUserByUsername: function(username, callback) {
+        connection.query('SELECT * FROM ' + this.tableName + ' WHERE username = ?', username, callback);
+    }
+};
+
+// Registration route
+route.post('/register', [
+    // Validate email and username fields
+    check('email').isEmail(),
+    check('username').isAlphanumeric().withMessage('Username must be alphanumeric'),
+
+    // Custom validation to check if email and username are unique
+    check('email').custom(async (value) => {
+        const user = await User.getUserByEmail(value);
+        if (user) {
+            throw new Error('Email already exists');
+        }
+    }),
+    check('username').custom(async (value) => {
+        const user = await User.getUserByUsername(value);
+        if (user) {
+            throw new Error('Username already exists');
+        }
+    }),
+], async (req, res) => {
+    // Check for validation errors
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(400).json({ errors: errors.array() });
+    }
+
+    // Hash the password
+    const saltRounds = 10;
+    const hashedPassword = await bcrypt.hash(req.body.password, saltRounds);
+
+    // Create a new user object
+    const newUser = {
+        email: req.body.email,
+        username: req.body.username,
+        password: hashedPassword,
+        full_name: req.body.full_name
+    };
+
+    // Insert user into MySQL
+    User.createUser(newUser, (error, results, fields) => {
+        if (error) {
+          console.error('Error inserting user: ' + error.message);
+          return res.status(500).json({ error: error.message });
+        }
+        console.log('Inserted a new user with id ' + results.insertId);
+        res.status(201).json(newUser);
+      });
+});
+
+// Login route
+route.post('/login', (req, res) => {
+    const { username, password } = req.body;
+    // Retrieve user from database
+    connection.query('SELECT * FROM users WHERE username = ?', [username], (err, results) => {
+        if (err) throw err;
+        if (results.length === 0) {
+            res.status(401).send('Invalid username or password');
+        } else {
+            const user = results[0];
+            // Compare passwords
+            bcrypt.compare(password, user.password, (err, isMatch) => {
+                if (err) throw err;
+                if (isMatch) {
+                    // Store user in session
+                    req.session.user = user;
+                    res.send('Login successful');
+                } else {
+                    res.status(401).send('Invalid username or password');
+                }
+            });
+        }
+    });
+});
+
+// Logout route
+route.post('/logout', (req, res) => {
+    req.session.destroy();
+    res.send('Logout successful');
+});
+
+//Dashboard route
+route.get('/dashboard', (req, res) => {
+    // Assuming you have middleware to handle user authentication and store user information in req.user
+    const userFullName = req.user.full_name;
+    res.render('dashboard', { fullName: userFullName });
+});
+
+// Route to retrieve course content
+route.get('/course/:id', (req, res) => {
+    const courseId = req.params.id;
+    const sql = 'SELECT * FROM courses WHERE id = ?';
+    db.query(sql, [courseId], (err, result) => {
+      if (err) {
+        throw err;
+      }
+      // Send course content as JSON response
+      res.json(result);
+    });
+});
+
+
+module.exports = route;
